サインアウトのhooks化
