+++ src/components/header/index.tsx
@@ -1,7 +1,9 @@
 import { Link } from "react-router-dom";
 import ReactLogo from "@/assets/react.svg";
+import { useSignout } from "@/modules/auth/use-signout";
 
 const Header = () => {
+  const { signout } = useSignout();
   return (
     <div className="w-full h-16 bg-white shadow-md fixed top-0 flex items-center pl-3 sm:px-6 z-50">
       <div className="flex w-full justify-between">
@@ -15,11 +17,11 @@ const Header = () => {
         <div className="flex sm:hidden items-center pr-6 space-x-4">
 
           {/* ログアウトボタン */}
-          <Link to="/signin">
-            <button className="px-2 py-1 text-sm text-gray-800 border border-gray-400 rounded hover:bg-gray-400">
-              ログアウト
-            </button>
-          </Link>
+          <button
+            onClick={signout}
+            className="px-2 py-1 text-sm text-gray-800 border border-gray-400 rounded hover:bg-gray-400">
+            ログアウト
+          </button>
 
           {/* 顧客登録ボタン */}
           <Link to="/create">
@@ -27,7 +29,7 @@ const Header = () => {
               ＋顧客登録
             </button>
           </Link>
-          
+
         </div>
       </div>
     </div>
